Separating the non-validation function to the file scanner
