+++ app/core/__init__.py
@@ -1,2 +1,2 @@
-from .validate import check_subdomain
 from .request import send_request, get_html_title
+from .scanner import check_subdomain
