+++ app/models/__init__.py
@@ -1,4 +1,3 @@
 from .scan_config import ScanConfig, current
 from .signatures import (HONEYPOT_NAME, HONEYPOT_TITLE, HONEYPOT_HASHES, HONEYPOT_SIZE_RANGE, HONEYPOT_SERVERS, OBSOLETE_VERSIONS, SUSPICIOUS_HEADER_ORDERS, TITLE_IGNORE,
-    CLOUDFLARE_IPS, CONFIDENCE_LABELS, SIGNAL_TIER, SIGNAL_WEIGHTS, USER_AGENT_FALLBACK)
-    CLOUDFLARE_IPS, CONFIDENCE_LABELS, SIGNAL_TIER, SIGNAL_WEIGHTS, HONEYPOT_HEADERS)
+    CLOUDFLARE_IPS, CONFIDENCE_LABELS, SIGNAL_TIER, SIGNAL_WEIGHTS, USER_AGENT_FALLBACK, HONEYPOT_HEADERS)
